rework file structure
