--- app/src/components/MainApp/MainApp.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default () => {
-    return <p>MainAPp</p>
-}
