--- app/src/components/hasAuth.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default (Login) => (App) => {
-    if(true){
-        return <App/>
-    }else{
-        return <Login/>
-    }
-}
