+++ app/src/components/hasAuth.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default Login => App => {
+  if (true) {
+    return <App />;
+  } else {
+    return <Login />;
+  }
+};
