+++ app/src/containers/App.jsx
@@ -3,8 +3,8 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import './App.css';
 import Login from './components/Login';
-import PrivateRoute from './components/private-route/PrivateRoute';
-import Home from './components/home/Home';
+import PrivateRoute from '../components/private-route/PrivateRoute';
+import Home from '../components/home/Home';
 
 class App extends Component {
   render() {
