+++ app/src/index.js
@@ -7,13 +7,15 @@ import App from './App';
 import './index.css';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
-import {rootReducer} from './reducers';
+import { rootReducer } from './reducers';
+
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 const store = createStore(
-  rootReducer, 
-  applyMiddleware(thunk, logger)
+  rootReducer,
+  composeEnhancers(applyMiddleware(thunk, logger))
 );
 
 ReactDOM.render(
